requirements.txt
feedparser
python-dateutil
tzdata